split reward managers
