+++ verl/utils/reward_score/__init__.py
@@ -11,20 +11,30 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from . import gsm8k, math, prime_math, prime_code
+# from . import gsm8k, math, prime_math, prime_code
 
 
 def _default_compute_score(data_source, solution_str, ground_truth):
     if data_source == 'openai/gsm8k':
-        return gsm8k.compute_score(solution_str, ground_truth)
+        from . import gsm8k
+        res = gsm8k.compute_score(solution_str, ground_truth)
     elif data_source in ['lighteval/MATH', 'DigitalLearningGmbH/MATH-lighteval']:
-        return math.compute_score(solution_str, ground_truth)
+        from . import math
+        res = math.compute_score(solution_str, ground_truth)
     elif data_source in [
             'numina_aops_forum', 'numina_synthetic_math', 'numina_amc_aime', 'numina_synthetic_amc', 'numina_cn_k12',
             'numina_olympiads'
     ]:
-        return prime_math.compute_score(solution_str, ground_truth)
+        from . import prime_math
+        res = prime_math.compute_score(solution_str, ground_truth)
     elif data_source in ['codecontests', 'apps', 'codeforces', 'taco']:
-        return prime_code.compute_score(solution_str, ground_truth, continuous=True)
+        from . import prime_code
+        res = prime_code.compute_score(solution_str, ground_truth, continuous=True)
     else:
         raise NotImplementedError
+
+    if isinstance(res, (int, float, bool)):
+        return float(res)
+    else:
+        return float(res[0])
+
